+++ edge-modules/azure-monitor/src/PrometheusMetricsParser.cs
@@ -5,6 +5,7 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
 {
     using System;
     using System.Collections.Generic;
+    using System.Globalization;
     using System.IO;
     using System.Linq;
     using System.Text.RegularExpressions;
@@ -17,7 +18,7 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
         //
         // - https://prometheus.io/docs/concepts/data_model/
         // - https://prometheus.io/docs/instrumenting/exposition_formats/
-        // - https://github.com/prometheus/common/blob/master/expfmt/text_parse.go
+        // - https://github.com/prometheus/common/blob/main/expfmt/text_parse.go
         //
         // Example:
         //
@@ -83,7 +84,7 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
                     }
 
                     double metricValue;
-                    if (!double.TryParse(match.Groups["metricvalue"]?.Value, out metricValue))
+                    if (!double.TryParse(match.Groups["metricvalue"]?.Value, NumberStyles.Float, NumberFormatInfo.InvariantInfo, out metricValue))
                     {
                         LoggerUtil.Writer.LogWarning($"Ignoring metric line because the metric value is malformed: [{line}]");
                         continue;