+++ edgelet/hsm-sys/azure-iot-hsm-c/src/hsm_key.h
@@ -32,11 +32,11 @@ MOCKABLE_FUNCTION(, KEY_HANDLE, create_encryption_key, const unsigned char*, key
 MOCKABLE_FUNCTION(, KEY_HANDLE, create_cert_key, const char*, key_file_name);
 
 MOCKABLE_FUNCTION(, int, generate_pki_cert_and_key, CERT_PROPS_HANDLE, cert_props_handle,
-                    int, serial_number, int, ca_path_len,
+                    int, ca_path_len,
                     const char*, key_file_name, const char*, cert_file_name,
                     const char*, issuer_key_file, const char*, issuer_certificate_file);
 MOCKABLE_FUNCTION(, int, generate_pki_cert_and_key_with_props, CERT_PROPS_HANDLE, cert_props_handle,
-                    int, serial_number, int, ca_path_len,
+                    long, serial_number, int, ca_path_len,
                     const char*, key_file_name, const char*, cert_file_name,
                     const PKI_KEY_PROPS*, key_props);
 MOCKABLE_FUNCTION(, int, generate_encryption_key, unsigned char**, key, size_t*, key_size);